feat: implement identity v2

+++ src/spaceone/identity/service/role_service.py
@@ -14,59 +14,60 @@ from spaceone.identity.manager.domain_manager import DomainManager
 _LOGGER = logging.getLogger(__name__)
 
 
+@authentication_handler
+@authorization_handler
+@mutation_handler
+@event_handler
 class RoleService(BaseService):
-
-    service = "identity"
     resource = "Role"
-    permission_group = "DOMAIN"
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.role_mgr = RoleManager()
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Role.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def create(self, params: RoleCreateRequest) -> Union[RoleResponse, dict]:
-        """ create role
+        """create role
 
          Args:
             params (RoleCreateRequest): {
                 'name': 'str',                          # required
                 'role_type': 'list',                    # required
-                'api_permissions': 'list',              # required
-                'page_permissions': 'list',
+                'permissions': 'list',                  # required
+                'page_access': 'list',
                 'tags': 'dict',
-                'domain_id': 'str'                      # required
+                'domain_id': 'str'                      # injected from auth
             }
 
         Returns:
             RoleResponse:
         """
 
-        if params.role_type in ['SYSTEM', 'SYSTEM_ADMIN']:
+        if params.role_type in ["SYSTEM", "SYSTEM_ADMIN"]:
             domain_mgr = DomainManager()
             domain_vo = domain_mgr.get_domain(params.domain_id)
 
-            root_domain_name = config.get_global('ROOT_DOMAIN_NAME', 'root')
+            root_domain_name = config.get_global("ROOT_DOMAIN_NAME", "root")
             if domain_vo.name != root_domain_name:
                 raise ERROR_PERMISSION_DENIED()
 
         role_vo = self.role_mgr.create_role(params.dict())
         return RoleResponse(**role_vo.to_dict())
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Role.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def update(self, params: RoleUpdateRequest) -> Union[RoleResponse, dict]:
-        """ update role
+        """update role
 
          Args:
             params (RoleUpdateRequest): {
                 'role_id': 'str',                       # required
                 'name': 'str',
-                'api_permissions': 'list',
-                'page_permissions': 'list',
+                'permissions': 'list',
+                'page_access': 'list',
                 'tags': 'dict',
-                'domain_id': 'str'                      # required
+                'domain_id': 'str'                      # injected from auth
             }
 
         Returns:
@@ -74,10 +75,8 @@ class RoleService(BaseService):
         """
 
         role_vo = self.role_mgr.get_role(params.role_id, params.domain_id)
-
-        if params.api_permissions:
-            # Check API Permissions
-            pass
+        if role_vo.is_managed:
+            raise ERROR_PERMISSION_DENIED()
 
         role_vo = self.role_mgr.update_role_by_vo(
             params.dict(exclude_unset=True), role_vo
@@ -85,15 +84,15 @@ class RoleService(BaseService):
 
         return RoleResponse(**role_vo.to_dict())
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Role.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def delete(self, params: RoleDeleteRequest) -> None:
-        """ delete role
+        """delete role
 
          Args:
             params (RoleDeleteRequest): {
                 'role_id': 'str',               # required
-                'domain_id': 'str',             # required
+                'domain_id': 'str',             # injected from auth
             }
 
         Returns:
@@ -101,17 +100,23 @@ class RoleService(BaseService):
         """
 
         role_vo = self.role_mgr.get_role(params.role_id, params.domain_id)
+        if role_vo.is_managed:
+            raise ERROR_PERMISSION_DENIED()
+        
         self.role_mgr.delete_role_by_vo(role_vo)
 
-    @transaction(scope="workspace_member:read")
+    @transaction(
+        permission="identity:Role.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
     @convert_model
     def get(self, params: RoleGetRequest) -> Union[RoleResponse, dict]:
-        """ get role
+        """get role
 
          Args:
             params (RoleGetRequest): {
                 'role_id': 'str',               # required
-                'domain_id': 'str',             # required
+                'domain_id': 'str',             # injected from auth
             }
 
         Returns:
@@ -121,19 +126,23 @@ class RoleService(BaseService):
         role_vo = self.role_mgr.get_role(params.role_id, params.domain_id)
         return RoleResponse(**role_vo.to_dict())
 
-    @transaction(scope="workspace_member:read")
-    @append_query_filter(['role_id', 'role_type', 'domain_id'])
-    @append_keyword_filter(['role_id', 'name'])
+    @transaction(
+        permission="identity:Role.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
+    @append_query_filter(["role_id", "name", "role_type", "domain_id"])
+    @append_keyword_filter(["role_id", "name"])
     @convert_model
     def list(self, params: RoleSearchQueryRequest) -> Union[RolesResponse, dict]:
-        """ list roles
+        """list roles
 
         Args:
             params (RoleSearchQueryRequest): {
                 'query': 'dict (spaceone.api.core.v1.Query)',
                 'role_id': 'str',
+                'name': 'str',
                 'role_type': 'str',
-                'domain_id': 'str',                     # required
+                'domain_id': 'str',                     # injected from auth
             }
 
         Returns:
@@ -146,17 +155,20 @@ class RoleService(BaseService):
         roles_info = [role_vo.to_dict() for role_vo in role_vos]
         return RolesResponse(results=roles_info, total_count=total_count)
 
-    @transaction(scope="workspace_member:read")
-    @append_query_filter(['domain_id'])
-    @append_keyword_filter(['role_id', 'name'])
+    @transaction(
+        permission="identity:Role.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
+    @append_query_filter(["domain_id"])
+    @append_keyword_filter(["role_id", "name"])
     @convert_model
     def stat(self, params: RoleStatQueryRequest) -> dict:
-        """ stat roles
+        """stat roles
 
         Args:
             params (PolicyStatQueryRequest): {
                 'query': 'dict (spaceone.api.core.v1.StatisticsQuery)', # required
-                'domain_id': 'str',         # required
+                'domain_id': 'str',         # injected from auth
             }
 
         Returns:
